+++ src/commands/mine.ts
